+++ src/applications/basic_analytics/stb_analytics/display_report_driver.py
